# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "description": "Documentação da API grades (notas), proposto no Bootcamp FullStack IGTI onde é possivel manipular as notas cadastradas no banco de dados.",
        "version": "1.0.0",
        "title": "API Grades IGTI",
        "contact": {
            "email": "ana.luhmaximiano@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/anamaximiano93/grade-api-igti/blob/main/LICENSE"
        }
    },
    "host": "http://localhost:8081/api-docs",
    "basePath": "/",
    "tags": [
        {
            "name": "Grades",
            "description": "Todas as rotas da aplicação"
        }
    ],
    "paths": {
        "/grade": {
            "post": {
                "tags": [
                    "Grades"
                ],
                "summary": "Cadastrar uma nova Grade.",
                "operationId": "createGrade",
                "requestBody": {
                    "description": "Obejeto necessario para cadastro.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Grade"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {}
                    },
                    "500": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Lista todas as Grades",
                "description": "Retorna uma lista de grade",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nome do aluno para retorno da grade",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GradesList"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Grades"
                ],
                "summary": "Deleta todas as Grades",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {}
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "deprecated": true
            }
        },
        "/grade/{id}": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Encontra uma grade",
                "description": "Retorna apenas uma grade",
                "operationId": "getGradeById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de retorno da grade",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Grade"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Grades"
                ],
                "summary": "Atualiza uma Grade",
                "description": "Atualiza apenas uma grade",
                "operationId": "updateGrade",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID para atualização da grade",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Objeto Grade Atualizado",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Grade"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {}
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Grades"
                ],
                "summary": "Deletar uma Grades",
                "description": "Deletar apenas uma grade",
                "operationId": "deleteGrade",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID para delete da grade",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {}
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "GradesList": {
                "type": "object",
                "properties": {
                    "count": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "xml": {
                    "name": "Grades"
                }
            },
            "Grade": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "subject": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "value": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "xml": {
                    "name": "Grades"
                }
            },
            "ApiResponse": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "type": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "externalDocs": {
        "description": "Find out more about Swagger",
        "url": "http://swagger.io"
    }
}